Refactored api controllers and api services
* Added more alternative response codes such as BadRequest and NoContent
* Changed naming to be more appropriate
* Made AddUserToMeeting to be httpPatch and moved it into MeetingController
* AddGoogleCalendarToUser is now httpPost
* Made angular api requesting services readable

+++ MeetingDateProposer/MeetingDateProposer/Controllers/MeetingController.cs
@@ -19,50 +19,92 @@ namespace MeetingDateProposer.Controllers
     public class MeetingController : ControllerBase
     {
         private readonly IMeetingService _meetingService;
+        private readonly IUserService _userService;
         private readonly IMapper _mapper;
 
-        public MeetingController(IMeetingService meetingService, IMapper mapper)
+        public MeetingController(
+            IMeetingService meetingService, 
+            IMapper mapper, 
+            IUserService userService)
         {
             _meetingService = meetingService;
             _mapper = mapper;
+            _userService = userService;
         }
 
-        [HttpGet("")]
+        [HttpGet]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status200OK)]
-        [ActionName("GetMeetingByIdAsync")]
+        [ActionName("GetMeetingById")]
         [AllowAnonymous]
-        public async Task<ActionResult<MeetingApiModel>> GetMeetingByIdAsync(Guid meetingId)
+        public async Task<ActionResult<MeetingApiModel>> GetMeetingById(Guid meetingId)
         {
-            var meeting = await _meetingService.GetMeetingByIdFromDbAsync(meetingId);
+            var meeting = await _meetingService.GetMeetingAsync(meetingId);
             if (meeting == null)
-            {
                 return NotFound();
-            }
 
-            var meetingViewModel = _mapper.Map<MeetingApiModel>(meeting);
-            return Ok(meetingViewModel);
+            var meetingApiModel = _mapper.Map<MeetingApiModel>(meeting);
+
+            return Ok(meetingApiModel);
         }
 
-        [HttpPost("")]
+        [HttpPost]
         [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [AllowAnonymous]
-        public async Task<ActionResult<MeetingApiModel>> CreateMeetingAsync(
+        public async Task<ActionResult<MeetingApiModel>> CreateMeeting(
             MeetingApiModel meetingApiModel)
         {
             var meeting = _mapper.Map<Meeting>(meetingApiModel);
-            await _meetingService.AddMeetingToDbAsync(meeting);
-            var meetingViewModel = _mapper.Map<MeetingApiModel>(meeting);
-            return CreatedAtAction(nameof(GetMeetingByIdAsync), new { meetingId = meeting.Id }, meetingViewModel);
+            await _meetingService.CreateMeetingAsync(meeting);
+            meetingApiModel = _mapper.Map<MeetingApiModel>(meeting);
+
+            return CreatedAtAction(nameof(GetMeetingById), new { meetingId = meeting.Id }, meetingApiModel);
+        }
+
+        [HttpPatch]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [AllowAnonymous]
+        public async Task<ActionResult<MeetingApiModel>> AddUserToMeeting(
+            Guid meetingId,
+            ApplicationUserApiModel applicationUserApiModel)
+        {
+            var meeting = await _meetingService.GetMeetingAsync(meetingId);
+            var userFromBody = _mapper.Map<ApplicationUser>(applicationUserApiModel);
+            var user = await _userService.GetUserAsync(userFromBody.Id);
+
+            if (meeting == null)
+                return NotFound();
+
+            if (meeting.ConnectedUsers.Exists(u => u.Id == userFromBody.Id))
+                return BadRequest();
+
+            if (user == null)
+            {
+                await _userService.CreateUserAsync(userFromBody);
+                user = userFromBody;
+            }
+
+            await _meetingService.AddUserToMeetingAsync(user, meeting);
+            var meetingApiModel = _mapper.Map<MeetingApiModel>(meeting);
+
+            return Ok(meetingApiModel);
         }
 
         [HttpDelete]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        public async Task<ActionResult> DeleteMeetingAsync(Guid meetingId)
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        public async Task<ActionResult> DeleteMeeting(Guid meetingId)
         {
+            var meeting = await _meetingService.GetMeetingAsync(meetingId);
+            if (meeting == null)
+                return NotFound();
+
             await _meetingService.DeleteMeetingAsync(meetingId);
-            return Ok();
+
+            return NoContent();
         }
     }
 }
